[feature](nereids) refactor statistics framework and introduce StatsCalculatorV2 (#12987)
* squash

change data type of metrics to double

unit test

add stats for some function

add stats for arithmeticExpr

1. set max/min of ColumnStats to double
2. add stats for binaryExpr/compoundExpr

in predicate

* Add LiteralExpr in ColumnStat just for user display only.

+++ fe/fe-core/src/main/java/org/apache/doris/analysis/SlotDescriptor.java
@@ -219,7 +219,7 @@ public class SlotDescriptor {
             }
         }
         // FIXME(dhc): mock ndv
-        stats.setNumDistinctValues(parent.getCardinality());
+        stats.setNumDistinctValues((long) parent.getCardinality());
         return stats;
     }
